Replace AvatarCreatorMenu with AvatarCreatorMenu2
